requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Data
yfinance>=0.2.28
requests>=2.31.0

# IBKR
ib_insync>=0.9.86

# FRED / Macro
fredapi>=0.5.1

# ML / LLM (optional)
openai>=1.12.0
anthropic>=0.18.0

# Telegram
python-telegram-bot>=20.7

# Config / Utils
python-dotenv>=1.0.0
pyyaml>=6.0.1
schedule>=1.2.1

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0

# Backtesting
backtrader>=1.9.78.123

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0
matplotlib>=3.8.0
seaborn>=0.13.0
